# openapi.v2.json
{
  "swagger": "2.0",
  "info": {
    "title": "SYS API Document",
    "version": "1.0",
    "description": "SYS API Document"
  },
  "host": "localhost:3000",
  "schemes": ["http", "https"],
  "produces": ["application/json"],
  "consumes": ["application/json"],
  "paths": {
...
    "/trd/add-retail": {
      "post": {
        "summary": "Create New Retail",
        "operationId": "post-trd-add-retail",
        "tags": ["Trd"],
        "description": "เพิ่มรายชื่อร้านค้า",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ret_id": {
                  "type": "integer",
                  "description": "ไอดีร้านค้าย่อย"
                },
                "trd_id": {
                  "type": "integer",
                  "description": "ไอดีผู้ขายเศษเหล็ก"
                }
              },
              "required": ["ret_id", "trd_id"]
            }
          }
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Successfully"
          }
        }
      }
    },
    "/trd/confirm": {
      "post": {
        "summary": "POST Confirm Retail",
        "operationId": "post-trd-confirm",
        "description": "ยืนยันร้านค้า",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "trd_id": {
                  "type": "integer",
                  "description": "ไอดีผู้ขายเศษเหล็ก"
                },
                "ret_id": {
                  "type": "integer",
                  "description": "ไอดีร้านค้า"
                }
              },
              "required": ["trd_id", "ret_id"]
            }
          }
        ],
        "tags": ["Trd"],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Successfully"
          }
        }
      }
    },
...
    "/trd/profile": {
      "get": {
        "summary": "GET Trd Profile",
        "tags": ["Trd"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Trd"
            },
            "examples": {
              "example-1": {
                "trd_id": "1",
                "trd_code": "trd001",
                "trd_title": "บจก.",
                "trd_name": "พิรุณสร้าง",
                "trd_adr1": "21/12 ซ.xxxx ",
                "trd_adr2": "ต.xxxx อ.xxxxx",
                "trdsap_code": "",
                "prov_code": "1000",
                "zip_code": "100000",
                "cn_p1": "บรรจง สมคิด",
                "cn_p1_tel": "081 3300022",
                "cn_p2": "",
                "cn_p2_tel": "",
                "cn_p3": "",
                "cn_p3_tel": "",
                "fax1": "",
                "fax2": "",
                "trd_email": "",
                "trd_profile": "",
                "trd_status_id": "1",
                "user_id": "",
                "created_by": "1",
                "created_at": "27/1/2021 21:32:48",
                "updated_by": "",
                "updated_at": ""
              }
            }
          }
        },
        "operationId": "get-trd-profile",
        "description": "ข้อมูลผู้ขายเศษเหล็ก",
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/trd/create": {
      "post": {
        "summary": "POST Create New Trd",
        "operationId": "post-trd-create",
        "responses": {
          "201": {
            "$ref": "#/responses/Successfully"
          }
        },
        "description": "สร้างรายการผู้ขาย",
        "security": [
          {
            "JWT": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Trd"
            }
          }
        ],
        "tags": ["Trd"]
      }
    },
    "/trd/update": {
      "post": {
        "summary": "POST Update Trd",
        "operationId": "post-trd-update",
        "responses": {
          "200": {
            "$ref": "#/responses/Successfully"
          }
        },
        "tags": ["Trd"],
        "description": "แก้ไขรายการ",
        "security": [
          {
            "JWT": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Trd"
            },
            "description": ""
          },
          {
            "type": "integer",
            "in": "query",
            "name": "trd_id",
            "description": "ไอดี",
            "required": true
          }
        ]
      }
    },
    "/trd/del": {
      "delete": {
        "summary": "DELETE Trd",
        "operationId": "delete-trd-del",
        "responses": {
          "200": {
            "$ref": "#/responses/Deleted"
          }
        },
        "tags": ["Trd"],
        "description": "ลบรายการ",
        "parameters": [
          {
            "type": "integer",
            "in": "query",
            "name": "trd_id",
            "description": "ไอดีผู้ขาย",
            "required": true
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Trd": {
      "title": "Trd",
      "type": "object",
      "description": "ผู้ขาย",
      "x-tags": ["Trd"],
      "properties": {
        "trd_id": {
          "type": "integer",
          "description": "autoid"
        },
        "trd_code": {
          "type": "string",
          "description": "รหัสผู้ขายเศษเหล็ก",
          "maxLength": 10,
          "minLength": 0
        },
        "trd_title": {
          "type": "string",
          "description": "ชื่อต้น",
          "maxLength": 10,
          "minLength": 0
        },
        "trd_name": {
          "type": "string",
          "description": "ชื่อผู้ขายเศษเหล็ก"
        },
        "trd_adr1": {
          "type": "string",
          "description": "ที่อยู่ 1"
        },
        "trd_adr2": {
          "type": "string",
          "description": "ที่อยู่ 2"
        },
        "trdsap_code": {
          "type": "string"
        },
        "prov_id": {
          "type": "integer",
          "description": "รหัสจังหวัด"
        },
        "zip_code": {
          "type": "string",
          "description": "รหัสไปรษณีย์"
        },
        "cn_p1": {
          "type": "string",
          "description": "ชื่อผู้ติดต่อ 1"
        },
        "cn_p1_tel": {
          "type": "string",
          "description": "เบอร์โทรผู้ติดต่อ1"
        },
        "cn_p2": {
          "type": "string",
          "description": "ชื่อผู้ติดต่อ2"
        },
        "cn_p2_tel": {
          "type": "string",
          "description": "เบอร์โทรผู้ติดต่อ2"
        },
        "cn_p3": {
          "type": "string",
          "description": "ชื่อผู้ติดต่อ3"
        },
        "cn_p3_tel": {
          "type": "string",
          "description": "เบอร์โทรผู้ติดต่อ3"
        },
        "fax1": {
          "type": "string",
          "description": "Fax 1"
        },
        "fax2": {
          "type": "string",
          "description": "Fax 2"
        },
        "trd_email": {
          "type": "string",
          "description": "อีเมล"
        },
        "trd_profile": {
          "type": "string"
        },
        "trd_status_id": {
          "type": "integer",
          "description": "สถานะ 1 active 2 deactive"
        },
        "user_id": {
          "type": "integer",
          "description": "ไอดีผู้ใช้งาน"
        },
        "created_by": {
          "type": "integer",
          "description": "ผู้สร้างรายการ"
        },
        "created_at": {
          "type": "string",
          "description": "วัน,เวลาที่สร้าง",
          "format": "date-time",
          "example": "2020-01-01 20:00:00"
        },
        "updated_by": {
          "type": "integer",
          "description": "ผู้แก้ไข"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "example": "2020-01-01 20:00:00",
          "description": "วัน,เวลาที่แก้ไข"
        }
      },
      "required": ["trd_code", "trd_title", "trd_name", "created_by", "created_at", "updated_by", "updated_at"]
    },
...
  },
  "responses": {
...
    "Successfully": {
      "description": "Example response",
      "content": {
        "application/json": {
          "schema": {
            "properties": {
              "id": {
                "type": "string"
              }
            }
          }
        }
      },
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Successfully."
          }
        }
      },
      "examples": {
        "example-1": {
          "message": "Successfully."
        }
      }
    },
    "Deleted": {
      "description": "Example response",
      "content": {
        "application/json": {
          "schema": {
            "properties": {
              "id": {
                "type": "string"
              }
            }
          }
        }
      },
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Deleted."
          }
        }
      },
      "examples": {
        "example-1": {
          "message": "Deleted."
        }
      }
    },
...
  },
...
  "basePath": "/api/v1",
  "securityDefinitions": {
    "JWT": {
      "name": "Authorization",
      "type": "apiKey",
      "in": "header",
      "description": "Bearer <token>"
    }
  }
}